Governance 2 (Part 1, Non-Controversial Changes) (#5892)
* initial stuff

* make copy data from kusama

* Update runtime/kusama/src/governance/old.rs


* Update runtime/kusama/src/governance/mod.rs
