+++ xcm/pallet-xcm/src/lib.rs
@@ -1517,8 +1517,8 @@ where
 	}
 
 	#[cfg(feature = "runtime-benchmarks")]
-	fn successful_origin() -> O {
-		O::from(Origin::Xcm(Here.into()))
+	fn try_successful_origin() -> Result<O, ()> {
+		Ok(O::from(Origin::Xcm(Here.into())))
 	}
 }
 
@@ -1542,8 +1542,8 @@ where
 	}
 
 	#[cfg(feature = "runtime-benchmarks")]
-	fn successful_origin() -> O {
-		O::from(Origin::Response(Here.into()))
+	fn try_successful_origin() -> Result<O, ()> {
+		Ok(O::from(Origin::Response(Here.into())))
 	}
 }
